+++ CHANGELOG.rst
@@ -11,6 +11,9 @@ DEV
   are handled with the correct localization.
 * shop.util.loader.get_model_string now can be used to load models outside
   the shop's namespace.
+* In cart_modifiers, replaced tuple (label, value) against a new structure
+  ``ExtraEntryLine`` which contains an extra field ``data`` to hold custom data
+  to be forwarded into the template views for orders and invoices.
 
 Version 0.0.14
 ==============